requirements.txt
# Core AI
anthropic>=0.40.0

# WhatsApp Integration
twilio>=9.0.0

# Scraping
instaloader>=4.10.0
apify-client>=1.6.0

# RSS Feed Parsing
feedparser>=6.0.0

# Database
sqlalchemy>=2.0.0

# Scheduling
apscheduler>=3.10.0

# HTTP Requests
httpx>=0.27.0
aiohttp>=3.9.0

# Environment Variables
python-dotenv>=1.0.0

# Date/Time
python-dateutil>=2.8.0
pytz>=2024.1

# Data Processing
pandas>=2.0.0

# Utilities
pydantic>=2.0.0
rich>=13.0.0  # Beautiful console output
